refactor: move AddBatchToQueue class into common
