+++ application/CohortManager/src/Functions/Shared/Common/AddBatchToQueue.cs
@@ -1,22 +1,17 @@
-namespace NHS.Screening.ReceiveCaasFile;
+namespace Common;
 
 using System.Collections.Concurrent;
-using System.Text;
-using System.Text.Json;
-using Azure.Storage.Queues;
-using Common;
 using Microsoft.Extensions.Logging;
 using Model;
 
 public class AddBatchToQueue : IAddBatchToQueue
 {
-
     public readonly ILogger<AddBatchToQueue> _logger;
-
-
     private readonly IAzureQueueStorageHelper _queueHelper;
 
-    public AddBatchToQueue(ILogger<AddBatchToQueue> logger, IAzureQueueStorageHelper queueHelper)
+    public AddBatchToQueue(
+        ILogger<AddBatchToQueue> logger,
+        IAzureQueueStorageHelper queueHelper)
     {
         _logger = logger;
         _queueHelper = queueHelper;
@@ -35,15 +30,17 @@ public class AddBatchToQueue : IAddBatchToQueue
         var itemsToAdd = currentBatch;
 
         // List of tasks to handle messages
-        List<Task> tasks = new List<Task>();
-        tasks.Add(Task.Factory.StartNew(() =>
-        {
-            // Process messages while there are items in the queue
-            while (itemsToAdd.TryDequeue(out var item))
+        List<Task> tasks =
+        [
+            Task.Factory.StartNew(async () =>
             {
-                AddMessage(item, queueName);
-            }
-        }));
+                // Process messages while there are items in the queue
+                while (itemsToAdd.TryDequeue(out var item))
+                {
+                    await AddMessage(item, queueName);
+                }
+            }),
+        ];
 
         // Wait for all tasks to complete
         await Task.WhenAll(tasks.ToArray());
@@ -52,6 +49,6 @@ public class AddBatchToQueue : IAddBatchToQueue
 
     private async Task AddMessage(BasicParticipantCsvRecord basicParticipantCsvRecord, string queueName)
     {
-        await _queueHelper.AddItemToQueueAsync<BasicParticipantCsvRecord>(basicParticipantCsvRecord, queueName);
+        await _queueHelper.AddItemToQueueAsync(basicParticipantCsvRecord, queueName);
     }
 }
